Split `Time` component into `RodosTime` and `RealTime`
This is necessary to break the hopefully last dependency cycle `Time` →
`FramSections` → `Periphery` → `Time` into `RealTime` → `FramSections` →
`Perihery` → `RodosTime`.
