+++ CMakeLists.txt
@@ -33,9 +33,10 @@ add_library(Sts1CobcSw_FramSections INTERFACE)
 add_library(Sts1CobcSw_Hal STATIC)
 add_library(Sts1CobcSw_Outcome INTERFACE)
 add_library(Sts1CobcSw_Periphery STATIC)
+add_library(Sts1CobcSw_RealTime INTERFACE)
+add_library(Sts1CobcSw_RodosTime INTERFACE)
 add_library(Sts1CobcSw_Serial INTERFACE)
 add_library(Sts1CobcSw_Telemetry STATIC)
-add_library(Sts1CobcSw_Time INTERFACE)
 add_library(Sts1CobcSw_Utility STATIC)
 add_library(Sts1CobcSw_Vocabulary INTERFACE)
 add_program(HelloDummy)
