+++ Sts1CobcSw/CMakeLists.txt
@@ -5,9 +5,10 @@ add_subdirectory(FramSections)
 add_subdirectory(Hal)
 add_subdirectory(Outcome)
 add_subdirectory(Periphery)
+add_subdirectory(RealTime)
+add_subdirectory(RodosTime)
 add_subdirectory(Serial)
 add_subdirectory(Telemetry)
-add_subdirectory(Time)
 add_subdirectory(Utility)
 add_subdirectory(Vocabulary)
