+++ Sts1CobcSw/CobcSoftware/CommandParser.cpp
@@ -5,8 +5,8 @@
 #include <Sts1CobcSw/FramSections/FramLayout.hpp>
 #include <Sts1CobcSw/FramSections/FramVector.hpp>
 #include <Sts1CobcSw/FramSections/PersistentVariables.hpp>
+#include <Sts1CobcSw/RealTime/RealTime.hpp>
 #include <Sts1CobcSw/Serial/Serial.hpp>
-#include <Sts1CobcSw/Time/RealTime.hpp>
 #include <Sts1CobcSw/Utility/DebugPrint.hpp>
 
 #include <strong_type/type.hpp>
