+++ Sts1CobcSw/CobcSoftware/EduProgramQueueThread.cpp
@@ -11,8 +11,8 @@
 #include <Sts1CobcSw/FramSections/FramVector.hpp>
 #include <Sts1CobcSw/FramSections/PersistentVariables.hpp>
 #include <Sts1CobcSw/Outcome/Outcome.hpp>
-#include <Sts1CobcSw/Time/RealTime.hpp>
-#include <Sts1CobcSw/Time/RodosTime.hpp>
+#include <Sts1CobcSw/RealTime/RealTime.hpp>
+#include <Sts1CobcSw/RodosTime/RodosTime.hpp>
 #include <Sts1CobcSw/Utility/DebugPrint.hpp>
 #include <Sts1CobcSw/Vocabulary/ProgramId.hpp>  // IWYU pragma: keep
 #include <Sts1CobcSw/Vocabulary/Time.hpp>
