+++ Sts1CobcSw/Hal/HardwareSpi.cpp
@@ -1,5 +1,5 @@
 #include <Sts1CobcSw/Hal/HardwareSpi.hpp>
-#include <Sts1CobcSw/Time/RodosTime.hpp>
+#include <Sts1CobcSw/RodosTime/RodosTime.hpp>
 
 #include <strong_type/affine_point.hpp>
