+++ Sts1CobcSw/RealTime/RealTime.ipp
@@ -2,8 +2,8 @@
 
 
 #include <Sts1CobcSw/FramSections/FramLayout.hpp>
-#include <Sts1CobcSw/Time/RealTime.hpp>
-#include <Sts1CobcSw/Time/RodosTime.hpp>
+#include <Sts1CobcSw/RealTime/RealTime.hpp>
+#include <Sts1CobcSw/RodosTime/RodosTime.hpp>
 
 #include <strong_type/type.hpp>
