--- Sts1CobcSw/Time/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_link_libraries(
-    Sts1CobcSw_Time INTERFACE rodos::without_main_on_linux strong_type::strong_type
-                              Sts1CobcSw_FramSections Sts1CobcSw_Vocabulary
-)
