+++ Tests/ManualTests/CMakeLists.txt
@@ -21,10 +21,10 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 Sts1CobcSw_Edu
                 Sts1CobcSw_Hal
                 Sts1CobcSw_Outcome
+                Sts1CobcSw_RealTime
                 Sts1CobcSw_Serial
-                Sts1CobcSw_Time
-                Sts1CobcSwTests::HardwareSetup
                 Sts1CobcSw_Vocabulary
+                Sts1CobcSwTests::HardwareSetup
     )
 
     add_test_program(Eps)
@@ -46,7 +46,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 Sts1CobcSw_Hal
                 Sts1CobcSw_Periphery
                 Sts1CobcSw_Serial
-                Sts1CobcSw_Time
+                Sts1CobcSw_RodosTime
                 Sts1CobcSw_Utility
                 Sts1CobcSwTests::HardwareSetup
     )
@@ -67,7 +67,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 Sts1CobcSw_Hal
                 Sts1CobcSw_Periphery
                 Sts1CobcSw_Serial
-                Sts1CobcSw_Time
+                Sts1CobcSw_RodosTime
                 Sts1CobcSw_Utility
                 Sts1CobcSw_Vocabulary
                 Sts1CobcSwTests::HardwareSetup
@@ -86,7 +86,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 strong_type::strong_type
                 Sts1CobcSw_Hal
                 Sts1CobcSw_Serial
-                Sts1CobcSw_Time
+                Sts1CobcSw_RodosTime
                 Sts1CobcSw_Utility
                 Sts1CobcSwTests::HardwareSetup
     )
@@ -94,7 +94,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
     add_test_program(Watchdog)
     target_link_libraries(
         Sts1CobcSwTests_Watchdog
-        PRIVATE rodos::rodos strong_type::strong_type Sts1CobcSw_Hal Sts1CobcSw_Time
+        PRIVATE rodos::rodos strong_type::strong_type Sts1CobcSw_Hal Sts1CobcSw_RodosTime
                 Sts1CobcSw_Vocabulary Sts1CobcSwTests::HardwareSetupWithoutWatchdog
     )
 endif()
